requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-dotenv==0.19.0
pydantic==1.8.2
httpx==0.19.0
python-multipart==0.0.5
torch==1.9.0
torchvision==0.10.0
numpy==1.21.2
pandas==1.3.3
scikit-learn==0.24.2
Pillow==8.3.2
joblib==1.0.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==3.2.0
aiofiles==0.7.0
python-jose[cryptography]==3.3.0
requests==2.26.0 